refactor: 모임 관련 컴포넌트 리팩토링
