+++ src/app/gathering/[id]/page.tsx
@@ -1,9 +1,10 @@
 import Breadcrumbs from "@/components/common/Breadcrumb";
+import GatheringViewer from "@/components/gathering/read/detail/GatheringViewer";
 import GatheringRecommendationList from "@/components/gathering/read/GatheringRecommendationList";
-import GatheringViewerContainer from "@/containers/gathering/read/detail/GatheringViewerContainer";
 import { cookies } from "next/headers";
 import { redirect } from "next/navigation";
 import { NextResponse } from "next/server";
+
 interface PageProps {
   params: Promise<{ id: string }>;
 }
@@ -96,12 +97,9 @@ async function getGathering(id: number) {
 
 export async function generateMetadata(props: PageProps) {
   const params = await props.params;
-
-  const {
-    id
-  } = params;
-
+  const { id } = params;
   const postId = Number(id);
+
   if (postId <= 0 || !Number.isSafeInteger(postId)) {
     throw new Error("페이지를 찾을 수 없습니다.");
   }
@@ -114,12 +112,9 @@ export async function generateMetadata(props: PageProps) {
 
 export default async function Page(props: PageProps) {
   const params = await props.params;
-
-  const {
-    id
-  } = params;
-
+  const { id } = params;
   const postId = Number(id);
+
   if (postId <= 0 || !Number.isSafeInteger(postId)) {
     return NextResponse.json(
       { message: "페이지를 찾을 수 없습니다." },
@@ -129,14 +124,10 @@ export default async function Page(props: PageProps) {
   const data = await getGathering(postId);
 
   return (
-    <div
-      className={
-        "m-auto flex min-h-[calc(100vh-25rem)] w-full max-w-[60rem] flex-col pb-[2.5rem]"
-      }
-    >
+    <div className="m-auto flex min-h-[calc(100vh-25rem)] w-full max-w-[60rem] flex-col pb-[2.5rem]">
       {/* 제일 상단 */}
       <Breadcrumbs categories={categories} />
-      <GatheringViewerContainer postId={postId} data={data} />
+      <GatheringViewer postId={postId} data={data} />
       <GatheringRecommendationList data={data.gatheringRecommend} />
     </div>
   );
