+++ src/app/gathering/write/page.tsx
@@ -1,4 +1,4 @@
-import GatheringCreateContainer from "@/containers/gathering/createUpdate/GatheringCreateContainer";
+import GatheringCreateEditor from "@/components/gathering/write/GatheringCreateEditor";
 import { Metadata } from "next";
 
 export const metadata: Metadata = {
@@ -9,7 +9,7 @@ export const metadata: Metadata = {
 export default function Page() {
   return (
     <div className={"min-h-[calc(100vh-25rem)] w-full pb-[2rem]"}>
-      <GatheringCreateContainer />
+      <GatheringCreateEditor />
     </div>
   );
 }
