+++ src/components/common/GatheringItemHome.tsx
@@ -1,11 +1,11 @@
-import GatheringBookMarkContainer from "@/containers/gathering/read/GatheringBookmarkContainer";
-import GatheringLikeContainer from "@/containers/gathering/read/GatheringLikeContainer";
+import GatheringBookMark from "@/components/gathering/read/GatheringBookmark";
 import { GatheringRecommend } from "@/types/GatheringDto";
 import { convertNumberToShortForm } from "@/utils/convertNumberToShortForm";
 import { format } from "date-fns";
 import { ko } from "date-fns/locale";
 import Image from "next/image";
 import Link from "next/link";
+import GatheringLike from "../gathering/read/GatheringLike";
 
 interface IGatheringItemHome {
   data: GatheringRecommend;
@@ -35,7 +35,7 @@ const GatheringItemHome = ({ data }: IGatheringItemHome) => {
           >
             <span>{data.gatheringCategoryName}</span>
           </p>
-          <GatheringBookMarkContainer
+          <GatheringBookMark
             isBookMark={data.isBookMark}
             postId={data.gatheringId}
           />
@@ -148,9 +148,9 @@ const GatheringItemHome = ({ data }: IGatheringItemHome) => {
           </div>
           <div className="flex flex-row justify-between">
             <div className="flex flex-row items-center gap-3">
-              <GatheringLikeContainer
-                likes={data.likeCount}
-                isLike={data.isLike}
+              <GatheringLike
+                initialLikes={data.likeCount}
+                initialIsLike={data.isLike}
                 gatheringId={data.gatheringId}
               />
               <div className="flex flex-row items-center gap-1 text-gray2">
