+++ src/components/gathering/createUpdate/editor/GatheringEditorChattingLink.tsx
@@ -1,3 +1,5 @@
+"use client";
+
 import { useFormContext } from "react-hook-form";
 
 const GatheringEditorChattingLink = () => {
