+++ src/components/gathering/createUpdate/editor/GatheringEditorContent.tsx
@@ -1,10 +1,12 @@
+"use client";
+
 import { useFormContext } from "react-hook-form";
 
 const GatheringEditorContent = () => {
   const formContext = useFormContext();
 
   return (
-    <article className={"flex w-full flex-col"}>
+    <article className="flex w-full flex-col">
       <div className="relative w-full">
         <textarea
           className={`min-h-[17.5rem] w-full resize-none rounded-[1rem] p-6 outline outline-[1px] outline-offset-[-1px] ${
@@ -12,9 +14,7 @@ const GatheringEditorContent = () => {
               ? "outline-red-500"
               : "outline-[#E3E3E3]"
           }`}
-          placeholder={
-            "어떤 모임을 만들어볼까요? 모임 정보 및 목표를 작성해 새로운 솔리들과 함께해보세요."
-          }
+          placeholder="어떤 모임을 만들어볼까요? 모임 정보 및 목표를 작성해 새로운 솔리들과 함께해보세요."
           {...formContext.register("content")}
           onChange={(e) => {
             formContext.setValue("content", e.target.value);
@@ -22,7 +22,7 @@ const GatheringEditorContent = () => {
           }}
           maxLength={500}
         />
-        <div className={"flex w-full justify-end pt-[0.75rem]"}>
+        <div className="flex w-full justify-end pt-[0.75rem]">
           {formContext.formState.errors.content && (
             <span className="absolute left-4 text-sm text-red-500">
               {formContext.formState.errors.content.message as String}
@@ -38,4 +38,5 @@ const GatheringEditorContent = () => {
     </article>
   );
 };
+
 export default GatheringEditorContent;
