--- src/containers/gathering/createUpdate/editor/GatheringEditorCategoryContainer.tsx
@@ -1,37 +1,0 @@
-import GatheringEditorCategory from "@/components/gathering/createUpdate/editor/GatheringEditorCategory";
-import useModalState from "@/hooks/useModalState";
-import { useEffect, useState } from "react";
-
-interface ICategory {
-  id: number;
-  name: string;
-  childrenCategories: ICategory[];
-}
-
-const GatheringEditorCategoryContainer = () => {
-  const modalState = useModalState();
-  const [categoryList, setCategoryList] = useState<ICategory[]>([]);
-
-  useEffect(() => {
-    const fetchData = async () => {
-      try {
-        const response = await fetch("/api/gathering/category");
-        if (!response.ok) {
-          throw new Error("Network response was not ok");
-        }
-        const result = await response.json();
-        setCategoryList(result);
-      } catch (error) {}
-    };
-
-    fetchData();
-  }, []);
-
-  return (
-    <GatheringEditorCategory
-      modalState={modalState}
-      categoryList={categoryList}
-    />
-  );
-};
-export default GatheringEditorCategoryContainer;
