--- src/containers/gathering/createUpdate/editor/GatheringEditorChattingLinkContainer.tsx
@@ -1,7 +1,0 @@
-import GatheringEditorChattingLink from "@/components/gathering/createUpdate/editor/GatheringEditorChattingLink";
-
-const GatheringEditorChattingLinkContainer = () => {
-  return <GatheringEditorChattingLink />;
-};
-
-export default GatheringEditorChattingLinkContainer;
