--- src/containers/gathering/createUpdate/editor/GatheringEditorContentContainer.tsx
@@ -1,6 +1,0 @@
-import GatheringEditorContent from "@/components/gathering/createUpdate/editor/GatheringEditorContent";
-
-const GatheringEditorContentContainer = () => {
-  return <GatheringEditorContent />;
-};
-export default GatheringEditorContentContainer;
