--- src/containers/gathering/createUpdate/editor/GatheringEditorDeadlineContainer.tsx
@@ -1,9 +1,0 @@
-import GatheringEditorDeadline from "@/components/gathering/createUpdate/editor/GatheringEditorDeadline";
-import useModalState from "@/hooks/useModalState";
-
-const GatheringEditorDeadlineContainer = () => {
-  const modalState = useModalState();
-
-  return <GatheringEditorDeadline modalState={modalState} />;
-};
-export default GatheringEditorDeadlineContainer;
