--- src/containers/gathering/createUpdate/editor/GatheringEditorParticipantsFilterContainer.tsx
@@ -1,8 +1,0 @@
-import GatheringEditorParticipantsFilter from "@/components/gathering/createUpdate/editor/GatheringEditorParticipantsFilter";
-import useModalState from "@/hooks/useModalState";
-
-const GatheringEditorParticipantsFilterContainer = () => {
-  const modalState = useModalState();
-  return <GatheringEditorParticipantsFilter modalState={modalState} />;
-};
-export default GatheringEditorParticipantsFilterContainer;
