--- src/containers/gathering/createUpdate/editor/GatheringEditorPeriodContainer.tsx
@@ -1,9 +1,0 @@
-import GatheringEditorPeriod from "@/components/gathering/createUpdate/editor/GatheringEditorPeriod";
-import useModalState from "@/hooks/useModalState";
-
-const GatheringEditorPeriodContainer = () => {
-  const modalState = useModalState();
-
-  return <GatheringEditorPeriod modalState={modalState} />;
-};
-export default GatheringEditorPeriodContainer;
