--- src/containers/gathering/createUpdate/editor/GatheringEditorTimeContainer.tsx
@@ -1,9 +1,0 @@
-import GatheringEditorTime from "@/components/gathering/createUpdate/editor/GatheringEditorTime";
-import useModalState from "@/hooks/useModalState";
-
-const GatheringEditorTimeContainer = () => {
-  const modalState = useModalState();
-
-  return <GatheringEditorTime modalState={modalState} />;
-};
-export default GatheringEditorTimeContainer;
