--- src/containers/gathering/createUpdate/editor/GatheringEditorTitleContainer.tsx
@@ -1,6 +1,0 @@
-import GatheringEditorTitle from "@/components/gathering/createUpdate/editor/GatheringEditorTitle";
-
-const GatheringEditorTitleContainer = () => {
-  return <GatheringEditorTitle />;
-};
-export default GatheringEditorTitleContainer;
